chore: Move casing functions to casing folder and reorganize exports
