+++ src/index.ts
@@ -1,110 +1,107 @@
 // Native
 export type { CharAt } from './native/char-at.js'
+export { charAt } from './native/char-at.js'
 export type { Concat } from './native/concat.js'
+export { concat } from './native/concat.js'
 export type { EndsWith } from './native/ends-with.js'
+export { endsWith } from './native/ends-with.js'
 export type { Includes } from './native/includes.js'
+export { includes } from './native/includes.js'
 export type { Join } from './native/join.js'
+export { join } from './native/join.js'
 export type { Length } from './native/length.js'
+export { length } from './native/length.js'
 export type { PadEnd } from './native/pad-end.js'
+export { padEnd } from './native/pad-end.js'
 export type { PadStart } from './native/pad-start.js'
+export { padStart } from './native/pad-start.js'
 export type { Repeat } from './native/repeat.js'
+export { repeat } from './native/repeat.js'
 export type { Replace } from './native/replace.js'
+export { replace } from './native/replace.js'
 export type { ReplaceAll } from './native/replace-all.js'
+export { replaceAll } from './native/replace-all.js'
 export type { Slice } from './native/slice.js'
+export { slice } from './native/slice.js'
 export type { Split } from './native/split.js'
+export { split } from './native/split.js'
 export type { StartsWith } from './native/starts-with.js'
+export { startsWith } from './native/starts-with.js'
 export type { TrimStart } from './native/trim-start.js'
+export { trimStart } from './native/trim-start.js'
 export type { TrimEnd } from './native/trim-end.js'
+export { trimEnd } from './native/trim-end.js'
 export type { Trim } from './native/trim.js'
 
-export { charAt } from './native/char-at.js'
-export { concat } from './native/concat.js'
-export { endsWith } from './native/ends-with.js'
-export { includes } from './native/includes.js'
-export { join } from './native/join.js'
-export { length } from './native/length.js'
-export { padEnd } from './native/pad-end.js'
-export { padStart } from './native/pad-start.js'
-export { repeat } from './native/repeat.js'
-export { replace } from './native/replace.js'
-export { replaceAll } from './native/replace-all.js'
-export { slice } from './native/slice.js'
-export { split } from './native/split.js'
-export { startsWith } from './native/starts-with.js'
+export { trim } from './native/trim.js'
 export { toLowerCase } from './native/to-lower-case.js'
 export { toUpperCase } from './native/to-upper-case.js'
-export { trimStart } from './native/trim-start.js'
-export { trimEnd } from './native/trim-end.js'
-export { trim } from './native/trim.js'
 
 // Utils
 export type { Reverse } from './utils/reverse'
-export type { Truncate } from './utils/truncate.js'
-export type { Words } from './utils/words.js'
-
 export { reverse } from './utils/reverse'
+export type { Truncate } from './utils/truncate.js'
 export { truncate } from './utils/truncate.js'
+export type { Words } from './utils/words.js'
 export { words } from './utils/words.js'
 
 // Characters
 export type { IsLetter, IsLower, IsUpper } from './utils/characters/letters.js'
-export type { Digit, IsDigit } from './utils/characters/numbers.js'
+export type { IsDigit } from './utils/characters/numbers.js'
 export type { IsSpecial } from './utils/characters/special.js'
-export type { Separator, IsSeparator } from './utils/characters/separators.js'
+export type { IsSeparator } from './utils/characters/separators.js'
 
 // Word casing
 export type { CamelCase } from './utils/word-case/camel-case.js'
-export type { ConstantCase } from './utils/word-case/constant-case.js'
-export type { DelimiterCase } from './utils/word-case/delimiter-case.js'
-export type { KebabCase } from './utils/word-case/kebab-case.js'
-export type { PascalCase } from './utils/word-case/pascal-case.js'
-export type { SnakeCase } from './utils/word-case/snake-case.js'
-export type { TitleCase } from './utils/word-case/title-case.js'
-
-export { capitalize } from './utils/capitalize.js'
-export { lowerCase } from './utils/word-case/lower-case.js'
 export { camelCase, toCamelCase } from './utils/word-case/camel-case.js'
+export type { ConstantCase } from './utils/word-case/constant-case.js'
 export {
   constantCase,
   toConstantCase,
 } from './utils/word-case/constant-case.js'
+export type { DelimiterCase } from './utils/word-case/delimiter-case.js'
 export {
   delimiterCase,
   toDelimiterCase,
 } from './utils/word-case/delimiter-case.js'
+export type { KebabCase } from './utils/word-case/kebab-case.js'
 export { kebabCase, toKebabCase } from './utils/word-case/kebab-case.js'
+export type { PascalCase } from './utils/word-case/pascal-case.js'
 export { pascalCase, toPascalCase } from './utils/word-case/pascal-case.js'
+export type { SnakeCase } from './utils/word-case/snake-case.js'
 export { snakeCase, toSnakeCase } from './utils/word-case/snake-case.js'
+export type { TitleCase } from './utils/word-case/title-case.js'
 export { titleCase, toTitleCase } from './utils/word-case/title-case.js'
-export { uncapitalize } from './utils/uncapitalize.js'
+
+export { capitalize } from './utils/word-case/capitalize.js'
+export { lowerCase } from './utils/word-case/lower-case.js'
+export { uncapitalize } from './utils/word-case/uncapitalize.js'
 export { upperCase } from './utils/word-case/upper-case.js'
 
 // Object keys word casing
 export type { CamelKeys } from './utils/object-keys/camel-keys.js'
-export type { ConstantKeys } from './utils/object-keys/constant-keys.js'
-export type { DelimiterKeys } from './utils/object-keys/delimiter-keys.js'
-export type { KebabKeys } from './utils/object-keys/kebab-keys.js'
-export type { PascalKeys } from './utils/object-keys/pascal-keys.js'
-export type { SnakeKeys } from './utils/object-keys/snake-keys.js'
-
 export { camelKeys } from './utils/object-keys/camel-keys.js'
+export type { ConstantKeys } from './utils/object-keys/constant-keys.js'
 export { constantKeys } from './utils/object-keys/constant-keys.js'
+export type { DelimiterKeys } from './utils/object-keys/delimiter-keys.js'
 export { delimiterKeys } from './utils/object-keys/delimiter-keys.js'
+export type { KebabKeys } from './utils/object-keys/kebab-keys.js'
 export { kebabKeys } from './utils/object-keys/kebab-keys.js'
+export type { PascalKeys } from './utils/object-keys/pascal-keys.js'
 export { pascalKeys } from './utils/object-keys/pascal-keys.js'
+export type { SnakeKeys } from './utils/object-keys/snake-keys.js'
 export { snakeKeys } from './utils/object-keys/snake-keys.js'
 
 // Object keys word casing (deep)
 export type { DeepCamelKeys } from './utils/object-keys/deep-camel-keys.js'
-export type { DeepConstantKeys } from './utils/object-keys/deep-constant-keys.js'
-export type { DeepDelimiterKeys } from './utils/object-keys/deep-delimiter-keys.js'
-export type { DeepKebabKeys } from './utils/object-keys/deep-kebab-keys.js'
-export type { DeepPascalKeys } from './utils/object-keys/deep-pascal-keys.js'
-export type { DeepSnakeKeys } from './utils/object-keys/deep-snake-keys.js'
-
 export { deepCamelKeys } from './utils/object-keys/deep-camel-keys.js'
+export type { DeepConstantKeys } from './utils/object-keys/deep-constant-keys.js'
 export { deepConstantKeys } from './utils/object-keys/deep-constant-keys.js'
+export type { DeepDelimiterKeys } from './utils/object-keys/deep-delimiter-keys.js'
 export { deepDelimiterKeys } from './utils/object-keys/deep-delimiter-keys.js'
+export type { DeepKebabKeys } from './utils/object-keys/deep-kebab-keys.js'
 export { deepKebabKeys } from './utils/object-keys/deep-kebab-keys.js'
+export type { DeepPascalKeys } from './utils/object-keys/deep-pascal-keys.js'
 export { deepPascalKeys } from './utils/object-keys/deep-pascal-keys.js'
+export type { DeepSnakeKeys } from './utils/object-keys/deep-snake-keys.js'
 export { deepSnakeKeys } from './utils/object-keys/deep-snake-keys.js'
