+++ src/utils/word-case/capitalize.ts
@@ -1,7 +1,7 @@
-import { charAt } from '../native/char-at.js'
-import { join } from '../native/join.js'
-import { slice } from '../native/slice.js'
-import { toUpperCase } from '../native/to-upper-case.js'
+import { charAt } from '../../native/char-at.js'
+import { join } from '../../native/join.js'
+import { slice } from '../../native/slice.js'
+import { toUpperCase } from '../../native/to-upper-case.js'
 
 /**
  * Capitalizes the first letter of a string. This is a runtime counterpart of `Capitalize<T>` from `src/types.d.ts`.
